requirements.txt
contourpy==1.1.0
cycler==0.11.0
fonttools==4.42.0
freetype-py==2.4.0
imageio==2.31.1
importlib-resources==6.0.1
joblib==1.3.1
kiwisolver==1.4.4
libigl==2.4.1
matplotlib==3.7.2
networkx==3.1
numpy==1.25.2
packaging==23.1
Pillow==10.0.0
pyglet==2.0.9
PyOpenGL==3.1.0
pyparsing==3.0.9
pyrender==0.1.45
python-dateutil==2.8.2
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
sklearn==0.0.post7
threadpoolctl==3.2.0
trimesh==3.23.1
zipp==3.16.2
